pyproject.toml
packages = ["netmonmqtt"]

[tool.setuptools.package-data]
netmonmqtt = []

[project]
# This is a unique identifier for your project, typically in the format <org_name>.<package_name>
name = "netmonmqtt"
version = "0.1.0"
description = "Network monitor that publishes to MQTT"
readme = "README.md"  # Assuming you have a README.md file

# SPDX license identifier. You can find a list of valid identifiers at https://spdx.org/licenses/
license = {text = "GPLv3"}

# Add your name and email address
authors = [
    {name = "Bryan Wyatt", email = "brwyatt@gmail.com"},
]

# Specify the minimum Python version required
requires-python = ">=3.11"

# Classify your project with PyPI trove classifiers
# See https://pypi.org/classifiers/ for a complete list
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring",
]

# Dependencies
dependencies = [
    "dnspython>=2.7.0,<3.0.0",
    "paho-mqtt>=2.1.0,<3.0.0",
    "py-machineid>=0.7.0,<1.0.0",
    "pythonping>=1.1.0,<1.2.0",
    "pyyaml>=6.0.0,<7.0.0",
    "setuptools>=75.8.0,<76.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=7.0.0,<9.0.0",
    "pre-commit>=2.0.0,<5.0.0",
    "tox>=4.23.2,<5.0.0",
    "netmonmqtt[test]",
]
test = [
    "coverage>=7.6.9,<8.0.0",
    "pytest>=8.3.4,<9.0.0",
]

[project.urls]
Homepage = "https://github.com/brwyatt/NetMonMQTT"
"Bug Tracker" = "https://github.com/brwyatt/NetMonMQTT/issues"

# Entry points for console scripts and plugins
[project.entry-points."console_scripts"]
netmonmqtt = "netmonmqtt.cli:main"

[build-system]
# These are the assumed build requirements if not specified
requires = ["setuptools>=50.0.0,<76.0.0"]
build-backend = "setuptools.build_meta"  # Specifies the backend to use for building the distribution

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
